% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenexRaster.R
\name{getModelledValuesToMatrix}
\alias{getModelledValuesToMatrix}
\title{Get matrix with phenex modelled values}
\usage{
getModelledValuesToMatrix(ndviObj)
}
\arguments{
\item{ndviObj}{An \code{NDVI} object generated by \code{\link[phenex]{modelNDVI}}}
}
\value{
A matrix containing modelled values
}
\description{
Just an ancillary function to push data from \code{NDVI} objects generated by
\code{\link[phenex]{modelNDVI}} and construct a data matrix
}
